requirements.txt
pyspark
pandas
SQLAlchemy
Flask
requests
python-dotenv
boto3
